refactor(giga-router): split into validator + rpc-only sibling impls
Reviewer asked whether splitting the rpc-only logic into its own type
would make the diff cleaner. Yes: the in-place mode-flag branching in
giga_router.go is gone, and the rpc-only-specific code lives in a new
file that reads top-to-bottom on its own.

What changes
- `GigaRouter` becomes an interface (8 methods: Run, RunInboundConn,
  LastCommittedBlockNumber, MaxGasPerBlock, BlockByNumber, BlockByHash,
  EvmProxy, IsRPCOnly). Returned by NewGigaRouter, stored in router.go
  as utils.Option[GigaRouter].
- `gigaRouterCommon` (new) holds fields and methods that are identical
  between modes — cfg, key, data, service, poolOut, BlockByNumber,
  BlockByHash, runExecute, executeBlock, translateGlobalBlock, and the
  shared dialAndRunConn (parameterized on a runClient callback so each
  impl picks RunClient vs RunBlockSyncClient).
- `gigaValidatorRouter` (in giga_router.go) embeds *gigaRouterCommon
  and adds consensus, producer, poolIn, inboundRPCOnlyPermits cap,
  lastCommitQCRecv. Has its own Run / RunInboundConn / EvmProxy /
  LastCommittedBlockNumber / MaxGasPerBlock / IsRPCOnly.
- `gigaRPCOnlyRouter` (in new giga_router_rpc_only.go) embeds
  *gigaRouterCommon and is otherwise stateless. Has its own Run (no
  consensus/producer, runs runRPCOnlySubscriber instead of dial-all),
  RunInboundConn (rejects), EvmProxy (no self short-circuit), the
  read-path source for LastCommittedBlockNumber (data.NextBlock-1)
  and MaxGasPerBlock (genesis). runRPCOnlySubscriber + its constants
  (rpcOnlyHealthyConnDuration, rpcOnlyMaxBackoff) move here too.
- No code duplication between the two impls: the only validator-vs-
  rpc-only differences are the things that genuinely differ
  (consensus/producer presence, inbound cap, dial topology, EvmProxy
  short-circuit, MaxGas source, LastCommittedBlockNumber source).

Tests
- TestGigaRouter_RPCOnly moves to its own file giga_router_rpc_only_test.go.
  Validator-flavored tests (FinalizeBlocks, EvmProxy, InitChainCommit)
  stay in giga_router_test.go alongside the testApp helpers they share.
- TestGigaRouter_EvmProxy and TestRouter_GigaSetWhenConfigured each
  add a single type-assert (*gigaValidatorRouter) where they previously
  reached into the concrete *GigaRouter struct.

External call sites (router.go, rpc/core/{blocks,status,mempool}.go)
keep the same method names — no caller changes beyond the type switch
from *GigaRouter to GigaRouter.
