+++ sei-tendermint/internal/p2p/giga_router_test.go
@@ -397,9 +397,12 @@ func TestGigaRouter_FinalizeBlocks(t *testing.T) {
 		// where producers placed the test txs.
 		giga0, _ := routers[0].Giga().Get()
 		latest := giga0.LastCommittedBlockNumber()
+		// Reach into the concrete validator impl to compare against
+		// data.State directly; this is an internal same-package test.
+		giga0Val := giga0.(*gigaValidatorRouter)
 		for h := int64(1); h <= latest; h++ {
 			gbn := atypes.GlobalBlockNumber(h) //nolint:gosec // h is positive
-			gb, err := giga0.data.GlobalBlock(ctx, gbn)
+			gb, err := giga0Val.data.GlobalBlock(ctx, gbn)
 			if err != nil {
 				continue // pruned out of the retain window
 			}
@@ -447,7 +450,7 @@ func TestGigaRouter_EvmProxy(t *testing.T) {
 	require.NoError(t, genDoc.ValidateAndComplete())
 
 	txMempool := mempool.NewTxMempool(mempool.TestConfig(), proxy.New(newTestApp(), proxy.NopMetrics()), mempool.NopMetrics(), mempool.NopTxConstraintsFetcher)
-	router, err := NewGigaRouter(&GigaRouterConfig{
+	routerIface, err := NewGigaRouter(&GigaRouterConfig{
 		DialInterval:   time.Second,
 		ValidatorAddrs: addrs,
 		Consensus: &consensus.Config{
@@ -467,6 +470,7 @@ func TestGigaRouter_EvmProxy(t *testing.T) {
 		GenDoc:    genDoc,
 	}, nodeKeys[0])
 	require.NoError(t, err)
+	router := routerIface.(*gigaValidatorRouter)
 
 	localValidator := validatorKeys[0].Public()
 	localURL, ok := urlByValidator[localValidator]
@@ -505,111 +509,4 @@ func TestGigaRouter_EvmProxy(t *testing.T) {
 	}
 
 	require.Equal(t, expectedRemoteURLs, returnedRemoteURLs)
-}
-
-// TestGigaRouter_RPCOnly covers the construction shape of the non-validator
-// (rpc-only) GigaRouter: routing always picks a remote shard owner (no
-// local short-circuit because there is no validator key), data + service
-// are constructed but consensus/producer are not, and the read-path
-// passthrough methods source values from the local data.State + genesis
-// doc (no errRPCOnly-style sentinels). The end-to-end block-sync /
-// executeBlock behaviour is covered by the autobahn integration test
-// where a real validator cluster supplies finalized blocks; this unit
-// test only verifies the construction surface.
-func TestGigaRouter_RPCOnly(t *testing.T) {
-	rng := utils.TestRng()
-	_, validatorKeys := atypes.GenCommittee(rng, 5)
-	addrs := map[atypes.PublicKey]GigaNodeAddr{}
-	urlByValidator := map[atypes.PublicKey]*url.URL{}
-	for i, validatorKey := range validatorKeys {
-		nodeKey := makeKey(rng)
-		addr := GigaNodeAddr{
-			Key:      nodeKey.Public(),
-			HostPort: tcp.HostPort{Hostname: "127.0.0.1", Port: 26657},
-		}
-		// Leave one validator without an EVMRPC URL so we exercise the
-		// "shard owner has no proxy" path too.
-		if i < 4 {
-			rpcURL, err := url.Parse(fmt.Sprintf("http://validator-%d.example.com:8545", i))
-			require.NoError(t, err)
-			addr.EVMRPC = utils.Some(rpcURL)
-			urlByValidator[validatorKey.Public()] = rpcURL
-		}
-		addrs[validatorKey.Public()] = addr
-	}
-	cp := types.DefaultConsensusParams()
-	cp.Block.MaxGas = 12345
-	genDoc := &types.GenesisDoc{
-		ChainID:         "giga-router-rpc-only-test",
-		InitialHeight:   1,
-		AppState:        testAppStateJSON(rng),
-		ConsensusParams: cp,
-	}
-	require.NoError(t, genDoc.ValidateAndComplete())
-
-	app := newTestApp()
-	txMempool := mempool.NewTxMempool(mempool.TestConfig(), proxy.New(app, proxy.NopMetrics()), mempool.NopMetrics(), mempool.NopTxConstraintsFetcher)
-
-	// Construct with no validator key (rpc-only nodes don't have one) and no
-	// Producer config. Consensus is set to an empty struct so the data WAL
-	// reads PersistentStateDir = None (in-memory) — fine for the
-	// construction-shape check.
-	router, err := NewGigaRouter(&GigaRouterConfig{
-		DialInterval:   time.Second,
-		ValidatorAddrs: addrs,
-		Consensus:      &consensus.Config{},
-		TxMempool:      txMempool,
-		GenDoc:         genDoc,
-		RPCOnly:        true,
-	}, makeKey(rng))
-	require.NoError(t, err)
-
-	// Shape: rpc-only routers run data + a block-sync service (outbound only)
-	// to pull finalized blocks from committee members. The validator-only
-	// state (consensus, producer, inbound peer pool, LastCommitQC watch)
-	// is structurally absent — router.validator is None.
-	require.True(t, router.IsRPCOnly())
-	require.NotNil(t, router.data)
-	require.NotNil(t, router.service)
-	require.NotNil(t, router.poolOut)
-	require.False(t, router.validator.IsPresent())
-
-	// EvmProxy: for every sender, the rpc-only router resolves to the shard
-	// owner's URL when the owner has one configured, and (nil,false) when it
-	// doesn't. Crucially, no sender is ever proxied "to ourselves" — that
-	// short-circuit doesn't exist in rpc-only mode.
-	expectedRemoteURLs := map[string]struct{}{}
-	for _, rpcURL := range urlByValidator {
-		expectedRemoteURLs[rpcURL.String()] = struct{}{}
-	}
-	returnedRemoteURLs := map[string]struct{}{}
-	for range 200 {
-		sender := common.BytesToAddress(utils.GenBytes(rng, common.AddressLength))
-		shardValidator := router.data.Committee().EvmShard(sender)
-		expectedURL, hasURL := urlByValidator[shardValidator]
-		proxyURL, ok := router.EvmProxy(sender)
-		if hasURL {
-			require.True(t, ok)
-			require.Equal(t, expectedURL.String(), proxyURL.String())
-			returnedRemoteURLs[proxyURL.String()] = struct{}{}
-		} else {
-			require.False(t, ok)
-			require.Nil(t, proxyURL)
-		}
-	}
-	// Sanity: with 200 random senders mapped uniformly over 5 shards we
-	// expect to have hit every shard owner with an EVMRPC URL at least once.
-	require.Equal(t, expectedRemoteURLs, returnedRemoteURLs)
-
-	// Read-path methods source from local data.State + genesis doc — no
-	// sentinels. Before any block is pushed, LastCommittedBlockNumber is
-	// InitialHeight - 1 (0 for genesis at 1); MaxGasPerBlock is the genesis
-	// consensus param.
-	require.Equal(t, int64(0), router.LastCommittedBlockNumber())
-	require.Equal(t, int64(12345), router.MaxGasPerBlock())
-	// BlockByHash returns &ResultBlock{Block:nil} for an unknown hash, the
-	// same shape the validator path returns — no sentinel mode-check.
-	rb, err := router.BlockByHash(t.Context(), atypes.BlockHeaderHash{})
-	require.NoError(t, err)
-	require.Nil(t, rb.Block)
 }
