+++ sei-tendermint/internal/rpc/core/env.go
@@ -129,9 +129,9 @@ func validatePage(pagePtr *int, perPage, totalCount int) (int, error) {
 //	if r, ok := env.gigaRouter().Get(); ok {
 //	    // Autobahn path, r is the router
 //	}
-func (env *Environment) gigaRouter() utils.Option[*p2p.GigaRouter] {
+func (env *Environment) gigaRouter() utils.Option[p2p.GigaRouter] {
 	if env.Router == nil { // inspect mode
-		return utils.None[*p2p.GigaRouter]()
+		return utils.None[p2p.GigaRouter]()
 	}
 	return env.Router.Giga()
 }
